% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{calcReadStats,GbsrGenotypeData-method}
\alias{calcReadStats,GbsrGenotypeData-method}
\title{Calculate mean, standard deviation, and quantile values of reads per sample and per marker.}
\usage{
\S4method{calcReadStats}{GbsrGenotypeData}(object, target = "both", q = NULL, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{q}{A numeric value [0-1] to indicate quantile to obtain.}
}
\description{
This function calculates several summary statistics of read counts
per marker and per sample. Those values will be stored in the SnpAnnotaionDataSet slot
and the ScanAnnotationDataSet slot and obtained via getter functions, e.g.
[getMeanReadRef()] and [getQtileReadAlt()].
}
\details{
Read count data can be obtained from the "annotation/format/AD/data" node or the
"annotation/format/AD/filt.data" node of the GDS file with `node = "raw"` or
`node = "filt"`, respectively. The [setCallFilter()] function generate filtered
read count data in the "annotation/format/AD/filt.data" node which can be accessed as
mentioned above.
}
\examples{
gdata <- loadGDS("/path/to/GDS.gds")
gdata <- calcReadStats(gdata, q = 0.5)
mean_reference_read_depth <- getMeanReadRef(gdata, target = "snp")
median_reference_read_depth <- getQtileReadAlt(gdata, target = "snp", q = 0.5)
hist(gdata, stats = "mean_ref")
hist(gdata, stats = "qtile_alt", q = 0.5)

}
