% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{calcReadStats}
\alias{calcReadStats}
\alias{calcReadStats,GbsrGenotypeData-method}
\title{Calculate mean, standard deviation, and quantile values of
reads per sample and per marker.}
\usage{
calcReadStats(object, target = "both", q = NULL, ...)

\S4method{calcReadStats}{GbsrGenotypeData}(object, target = "both", q = NULL)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{target}{Either of "snp" and "scan".}

\item{q}{A numeric value [0-1] to indicate quantile to obtain.}

\item{...}{Unused.}
}
\value{
A \link{GbsrGenotypeData} object with read statistics information.
}
\description{
This function calculates several summary statistics of read counts
per marker and per sample. Those values will be stored
in the SnpAnnotaionDataFrame slot and the \link{ScanAnnotationDataFrame} slot
and obtained via getter functions, e.g.
\code{\link[=getMeanReadRef]{getMeanReadRef()}} and \code{\link[=getQtileReadAlt]{getQtileReadAlt()}}.
}
\details{
Read count data can be obtained from the "annotation/format/AD/data" node
or the "annotation/format/AD/filt.data" node of the GDS file
with \code{node = "raw"} or \code{node = "filt"}, respectively.
The \code{\link[=setCallFilter]{setCallFilter()}} function generate filtered read count data
in the "annotation/format/AD/filt.data" node which can be accessed as
mentioned above.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)

# Calculate means, standard deviations, quantiles of read counts
# per marker and per sample with or without standardization of 
# the counts and store them in the
# [SnpAnnotationDataFrame] and [ScanAnnotationDataFrame] objects 
# linked at the slots of the [GbsrGenotypeData] object.
gdata <- calcReadStats(gdata, q = 0.5)

# Get the means of reference allele read counts per marker.
mean_reference_read_depth <- getMeanReadRef(gdata, target = "snp")

# Get the 0.5 percentiles (medians) of the alternative allele
# read counts per marker.
median_reference_read_depth <- getQtileReadAlt(gdata,
                                               target = "snp",
                                               q = 0.5)
# Draw histograms of the means of reference allele read counts
# per sample and marker.                      
histGBSR(gdata, stats = "mean_ref")

# Draw histograms of the 0.5 percentiles (medians) of 
# alternative allele read counts per sample and marker.
histGBSR(gdata, stats = "qtile_alt", q = 0.5)

# Close the connection to the GDS file.
closeGDS(gdata)

}
