% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{closeGDS,GbsrGenotypeData-method}
\alias{closeGDS,GbsrGenotypeData-method}
\title{Close the connection to the GDS file}
\usage{
\S4method{closeGDS}{GbsrGenotypeData}(object, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}
}
\description{
Close the connection to the GDS file linked to the given GbsrGenotypeData object.
}
