% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{closeGDS}
\alias{closeGDS}
\alias{closeGDS,GbsrGenotypeData-method}
\title{Close the connection to the GDS file}
\usage{
closeGDS(object, ...)

\S4method{closeGDS}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\description{
Close the connection to the GDS file linked to the given GbsrGenotypeData object.
}
