% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{countGenotype,GbsrGenotypeData-method}
\alias{countGenotype,GbsrGenotypeData-method}
\title{Count genotype calls and alleles per sample and per marker.}
\usage{
\S4method{countGenotype}{GbsrGenotypeData}(object, target = "both", node = "", ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{node}{Either of "raw", "filt", and "cor". See details.}
}
\description{
This function calculates several summary statistics of genotype calls and alleles
per marker and per sample. Those values will be stored in the SnpAnnotaionDataSet slot
and the ScanAnnotationDataSet slot and obtained via getter functions, e.g.
[getCountGenoRef()], [getCountAlleleRef()], and [getMAF()].
}
\details{
#' Genotype call data can be obtained from the "genotype" node, the "filt.genotype"
node, or the "corrected.genotype" node of the GDS file with `node = "raw"`,
`node = "filt"`, or `node = "raw"`, respectively.
The [setCallFilter()] function generate filtered genotype call data in the
"filt.genotype" node which can be accessed as mentioned above. On the other hand, the
"corrected.genotype" node can be generated via the [clean()] function.
}
\examples{
gdata <- loadGDS("/path/to/GDS.gds")
gdata <- countGenotype(gdata)
sample_missing_rate <- getCountGenoMissing(gdata, target = "scan", prop = TRUE)
marker_minor_allele_freq <- getMAF(gdata, target = "snp")
hist(gdata, stats = "missing")

}
