% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{countRead,GbsrGenotypeData-method}
\alias{countRead,GbsrGenotypeData-method}
\title{Count reads per sample and per marker.}
\usage{
\S4method{countRead}{GbsrGenotypeData}(object, target = "both", node = "", ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{node}{Either of "raw" and "filt". See details.}
}
\description{
This function calculates several summary statistics of read counts
per marker and per sample. Those values will be stored in the SnpAnnotaionDataSet slot
and the ScanAnnotationDataSet slot and obtained via getter functions, e.g.
[getCountReadRef()] and [getCountReadAlt()].
}
\details{
Read count data can be obtained from the "annotation/format/AD/data" node or the
"annotation/format/AD/filt.data" node of the GDS file with `node = "raw"` or
`node = "filt"`, respectively. The [setCallFilter()] function generate filtered
read count data in the "annotation/format/AD/filt.data" node which can be accessed as
mentioned above.
}
\examples{
gdata <- loadGDS("/path/to/GDS.gds")
gdata <- countRead(gdata)
read_depth_per_marker <- getCountRead(gdata, target = "snp")
reference_read_freq <- getCountReadRef(gdata, target = "snp", prop = TRUE)
hist(gdata, stats = "ad_ref")

}
