% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{countRead}
\alias{countRead}
\alias{countRead,GbsrGenotypeData-method}
\title{Count reads per sample and per marker.}
\usage{
countRead(object, target = "both", node = "", ...)

\S4method{countRead}{GbsrGenotypeData}(object, target = "both", node = "")
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{node}{Either of "raw" and "filt". See details.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object with read count information.
}
\description{
This function calculates several summary statistics of read counts
per marker and per sample. Those values will be stored
in the SnpAnnotaionDataSet slot and the ScanAnnotationDataSet slot
and obtained via getter functions, e.g.
\code{\link[=getCountReadRef]{getCountReadRef()}} and \code{\link[=getCountReadAlt]{getCountReadAlt()}}.
}
\details{
Read count data can be obtained from the "annotation/format/AD/data" node
or the "annotation/format/AD/filt.data" node of the GDS file
with \code{node = "raw"} or \code{node = "filt"}, respectively.
The \code{\link[=setCallFilter]{setCallFilter()}} function generate filtered read count data
in the "annotation/format/AD/filt.data" node which can be accessed as
mentioned above.
}
\examples{
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
gdata <- countRead(gdata)
read_depth_per_marker <- getCountRead(gdata, target = "snp")
reference_read_freq <- getCountReadRef(gdata, target = "snp", prop = TRUE)
histGBSR(gdata, stats = "ad_ref")
closeGDS(gdata) # Close the connection to the GDS file

}
