% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData_v2.R
\name{estDosage}
\alias{estDosage}
\alias{estDosage,GbsrGenotypeData-method}
\title{Estimate dosage}
\usage{
estDosage(
  object,
  recomb_rate = 0.04,
  error_rate = 0.0025,
  n_threads = 1,
  smooth = TRUE,
  mindp = 10,
  ...
)

\S4method{estDosage}{GbsrGenotypeData}(object, recomb_rate, error_rate, n_threads, smooth, mindp)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{n_threads}{An integer value to specify the number of
threads used for the calculation. The default is 1 and if \code{n_threads = NULL},
automatically set half the number of available threads on the computer.}

\item{...}{Unused.}
}
\value{
A \link{GbsrGenotypeData} object.
}
\description{
Estimate dosage
}
\details{
Calculate smoothed read ratio per marker.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

gds <- estDosage(gds)

# Close the connection to the GDS file.
closeGDS(gds)

}
