% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{flipData,GbsrGenotypeData-method}
\alias{flipData,GbsrGenotypeData-method}
\title{Flip genotype, allele information, and allele read counts of the flipped SNP markers.}
\usage{
\S4method{flipData}{GbsrGenotypeData}(object, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}
}
\description{
Genotype data, allele information, and allele read counts to be the expected reference
allele as actually reference allele.
}
\details{
Flipped markers are markers where the alleles expected as reference allele are called as
alternative allele. If you specify two parents in the `parents` argument of
[setParents()] with `flip = TRUE`, `bi = TRUE`, and `homo = TRUE`, the alleles found
in the parent specified as the first element to the `parents` argument are supposed as
reference alleles of the markers. If the "expected" reference alleles are not actually
called as reference alleles but alternative alleles in the given data. [setParents()] will
automatically labels those markers "flipped". The SnpAnnotatoinDataSet slot sores this
information and accessible via [getFlipped()] which gives you a logical vector
indicating which markers are labeled as flipped `TRUE` or not flipped `FALSE`.
[haveFlipped()] just tells you whether the SnpAnnotatoinDataSet slot has
the information of flipped markers or not.
}
\seealso{
[setParents()], [getFlipped()], and [haveFlipped()].
}
