% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{getAlleleA,GbsrGenotypeData-method}
\alias{getAlleleA,GbsrGenotypeData-method}
\title{Obtain reference allele information of each SNP marker}
\usage{
\S4method{getAlleleA}{GbsrGenotypeData}(object, valid = TRUE, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}
}
\description{
This function returns reference alleles, either of A, T, G, and C, of SNP markers.
}
\details{
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
