% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getAlleleA}
\alias{getAlleleA}
\alias{getAlleleA,GbsrGenotypeData-method}
\title{Obtain reference allele information of each SNP marker}
\usage{
getAlleleA(object, valid = TRUE, ...)

\S4method{getAlleleA}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A character vector indicating the reference alleles.
}
\description{
This function returns reference alleles, either
of A, T, G, and C, of SNP markers.
}
\details{
If \code{valid = TRUE}, the chromosome information of markers
which are labeled \code{TRUE} in the ScanAnnotationDataSet slot will be
returned. \code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
gds_fn <- system.file("extdata", "simpop.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
getAlleleA(gdata)

}
