% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getAlleleB}
\alias{getAlleleB}
\alias{getAlleleB,GbsrGenotypeData-method}
\title{Obtain alternative allele information of each SNP marker}
\usage{
getAlleleB(object, valid = TRUE, ...)

\S4method{getAlleleB}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\description{
This function returns alternative alleles, either of A, T, G, and C, of SNP markers.
}
\details{
If \code{valid = TRUE}, the chromosome information of markers which are labeled \code{TRUE}
in the ScanAnnotationDataSet slot will be returned. \code{\link[=getValidSnp]{getValidSnp()}} tells you
which samples are valid.
}
