% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{getChromosome,GbsrGenotypeData-method}
\alias{getChromosome,GbsrGenotypeData-method}
\title{Obtain chromosome information of each SNP marker}
\usage{
\S4method{getChromosome}{GbsrGenotypeData}(object, valid = TRUE, levels = FALSE, name = FALSE, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}

\item{levels}{A logical value. See details.}

\item{name}{A logical value. See details.}
}
\description{
This function returns indexes or names of chromsomes of each SNP or just a
set of unique chromosome names.
}
\details{
A GDS file created via GBScleanR stores chromosome names as sequential integers
from 1 to N, where N is the number of chromosomes. This function returns those
indexes as default. If you need actual names of the chromosomes, set `name = TRUE`.
`levels = TRUE` gives you only unique chromosome names with length N.
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
