% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getCountAlleleAlt}
\alias{getCountAlleleAlt}
\alias{getCountAlleleAlt,GbsrGenotypeData-method}
\title{Obtain total alternative allele counts per SNP or per scan (sample)}
\usage{
getCountAlleleAlt(object, target = "snp", valid = TRUE, prop = FALSE, ...)

\S4method{getCountAlleleAlt}{GbsrGenotypeData}(object, target = "snp", valid = TRUE, prop = FALSE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{valid}{A logical value. See details.}

\item{prop}{A logical value whether to return values as proportions of total alternative allele counts to total non missing allele counts or not.}

\item{...}{Unused.}
}
\description{
Obtain total alternative allele counts per SNP or per scan (sample)
}
\details{
You need to execute [countGenotype()] to calculate sumaary statisticsto be
obtained via this function.
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
