% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{getCountGenoRef,GbsrGenotypeData-method}
\alias{getCountGenoRef,GbsrGenotypeData-method}
\title{Obtain total reference genotype counts per SNP or per scan (sample)}
\usage{
\S4method{getCountGenoRef}{GbsrGenotypeData}(object, target = "snp", valid = TRUE, prop = FALSE, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{valid}{A logical value. See details.}

\item{prop}{A logical value whether to return values as proportions of total reference genotype counts to total non missing genotype counts or not.}
}
\description{
Obtain total reference genotype counts per SNP or per scan (sample)
}
\details{
You need to execute [countGenotype()] to calculate sumaary statisticsto be
obtained via this function.
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
