% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getCountRead}
\alias{getCountRead}
\alias{getCountRead,GbsrGenotypeData-method}
\title{Obtain total read counts per SNP or per scan (sample)}
\usage{
getCountRead(object, target = "snp", valid = TRUE, ...)

\S4method{getCountRead}{GbsrGenotypeData}(object, target = "snp", valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\description{
Obtain total read counts per SNP or per scan (sample)
}
\details{
You need to execute \code{\link[=countRead]{countRead()}} to calculate sumaary statisticsto be
obtained via this function.
If \code{valid = TRUE}, the chromosome information of markers which are labeled \code{TRUE}
in the ScanAnnotationDataSet slot will be returned. \code{\link[=getValidSnp]{getValidSnp()}} tells you
which samples are valid.
}
