% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getFlipped}
\alias{getFlipped}
\alias{getFlipped,GbsrGenotypeData-method}
\title{Get a logical vector indicating flipped SNP markers.}
\usage{
getFlipped(object, valid = TRUE, ...)

\S4method{getFlipped}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value to specify flip alleles only of valid markers. see also[setSnpFilter()].}

\item{...}{Unused.}
}
\description{
Get a logical vector indicating flipped SNP markers.
}
\details{
Flipped markers are markers where the alleles expected as reference allele are called as
alternative allele. If you specify two parents in the `parents` argument of
[setParents()] with `flip = TRUE`, `bi = TRUE`, and `homo = TRUE`, the alleles found
in the parent specified as the first element to the `parents` argument are supposed as
reference alleles of the markers. If the "expected" reference alleles are not actually
called as reference alleles but alternative alleles in the given data. [setParents()] will
automatically labels those markers "flipped". The SnpAnnotatoinDataSet slot sores this
information and accessible via [getFlipped()] which gives you a logical vector
indicating which markers are labeled as flipped `TRUE` or not flipped `FALSE`.
[haveFlipped()] just tells you whether the SnpAnnotatoinDataSet slot has
the information of flipped markers or not.
}
\seealso{
[setParents()] and [haveFlipped()].
}
