% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{getInfo,GbsrGenotypeData-method}
\alias{getInfo,GbsrGenotypeData-method}
\title{Obtain information stored in the "annotation/info" node}
\usage{
\S4method{getInfo}{GbsrGenotypeData}(object, var, valid = TRUE, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{var}{A string to indicate which annotation info should be retrieved.}

\item{valid}{A logical value. See details.}
}
\description{
The "annotation/info" node stores annotation infromation of markers obtained
via SNP calling tools like bcftools and GATK.
}
\details{
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
