% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getMAC}
\alias{getMAC}
\alias{getMAC,GbsrGenotypeData-method}
\title{Obtain minor allele counts per SNP or per scan (sample)}
\usage{
getMAC(object, target = "snp", valid = TRUE, ...)

\S4method{getMAC}{GbsrGenotypeData}(object, target = "snp", valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\description{
Obtain minor allele counts per SNP or per scan (sample)
}
\details{
You need to execute [countGenotype()] to calculate sumaary statisticsto be
obtained via this function.
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
