% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{getMeanReadAlt,GbsrGenotypeData-method}
\alias{getMeanReadAlt,GbsrGenotypeData-method}
\title{Obtain mean values of total alternative read counts per SNP or per scan (sample)}
\usage{
\S4method{getMeanReadAlt}{GbsrGenotypeData}(object, target = "snp", valid = TRUE, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{valid}{A logical value. See details.}
}
\description{
Obtain mean values of total alternative read counts per SNP or per scan (sample)
}
\details{
You need to execute [calcReadStats()] to calculate sumaary statisticsto be
obtained via this function.
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
