% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{getPloidy,GbsrGenotypeData-method}
\alias{getPloidy,GbsrGenotypeData-method}
\title{Obtain ploidy information of each SNP marker}
\usage{
\S4method{getPloidy}{GbsrGenotypeData}(object, valid = TRUE, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}
}
\description{
This function returns ploidy of each SNP marker. The ploidy of all the markers in
a dataset is a same value and the current implementation of GBScleanR only works
with data having ploidy = 2 for all markers.
}
\details{
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
