% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getPloidy}
\alias{getPloidy}
\alias{getPloidy,GbsrGenotypeData-method}
\title{Obtain ploidy information of each SNP marker}
\usage{
getPloidy(object, valid = TRUE, ...)

\S4method{getPloidy}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A integer vector indicating the ploidy of each marker.
}
\description{
This function returns ploidy of each SNP marker. The ploidy of all
the markers in a dataset is a same value and the current implementation
of \link{GBScleanR} only works with data having ploidy = 2 for all markers.
}
\details{
If \code{valid = TRUE}, the chromosome information of markers which are
labeled \code{TRUE} in the \link{ScanAnnotationDataFrame} slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)

getPloidy(gdata)

# Close the connection to the GDS file.
closeGDS(gdata)

}
