% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getPosition}
\alias{getPosition}
\alias{getPosition,GbsrGenotypeData-method}
\title{Obtain physical position information of each SNP marker}
\usage{
getPosition(object, valid = TRUE, ...)

\S4method{getPosition}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\description{
This function returns physical positions of SNP markers.
}
\details{
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
