% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{getQtileReadAlt,GbsrGenotypeData-method}
\alias{getQtileReadAlt,GbsrGenotypeData-method}
\title{Obtain quantile values of total alternative read counts per SNP or per scan (sample)}
\usage{
\S4method{getQtileReadAlt}{GbsrGenotypeData}(object, target = "snp", q = 0.5, valid = TRUE, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{q}{A numeric value [0-1] to indicate quantile to obtain.}

\item{valid}{A logical value. See details.}
}
\description{
Obtain quantile values of total alternative read counts per SNP or per scan (sample)
}
\details{
You need to execute [calcReadStats()] to calculate sumaary statisticsto be
obtained via this function.
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
