% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getQtileReadRef}
\alias{getQtileReadRef}
\alias{getQtileReadRef,GbsrGenotypeData-method}
\title{Obtain quantile values of total reference read counts per
SNP or per scan (sample)}
\usage{
getQtileReadRef(object, target = "snp", q = 0.5, valid = TRUE, ...)

\S4method{getQtileReadRef}{GbsrGenotypeData}(object, target = "snp", q = 0.5, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{q}{A numeric value \link{0-1} to indicate quantile to obtain.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A numeric vector of the quantile values of
alternative allele reads per marker.
}
\description{
Obtain quantile values of total reference read counts per
SNP or per scan (sample)
}
\details{
You need to execute \code{\link[=calcReadStats]{calcReadStats()}} to calculate sumaary statisticsto be
obtained via this function.
If \code{valid = TRUE}, the chromosome information of markers which are
labeled \code{TRUE} in the ScanAnnotationDataSet slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
gds_fn <- system.file("extdata", "simpop.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
gdata <- calcReadStats(gdata)
getQtileReadRef(gdata)

}
