% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getScanID}
\alias{getScanID}
\alias{getScanID,GbsrGenotypeData-method}
\title{Obtain scan (sample) ID}
\usage{
getScanID(object, valid = TRUE, ...)

\S4method{getScanID}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\description{
This function returns scan (sample) ID.
}
\details{
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
