% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getSnpID}
\alias{getSnpID}
\alias{getSnpID,GbsrGenotypeData-method}
\title{Obtain SNP ID}
\usage{
getSnpID(object, valid = TRUE, ...)

\S4method{getSnpID}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A character vector of SNP IDs.
}
\description{
This function returns SNP ID of SNP markers.
}
\details{
If \code{valid = TRUE}, the chromosome information of markers which are
labeled \code{TRUE} in the \link{ScanAnnotationDataFrame} slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)

getSnpID(gdata)

# Close the connection to the GDS file.
closeGDS(gdata)

}
