% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{getValidScan,GbsrGenotypeData-method}
\alias{getValidScan,GbsrGenotypeData-method}
\title{Return a logical vector indicating which are valid scans (samples).}
\usage{
\S4method{getValidScan}{GbsrGenotypeData}(object, parents = FALSE, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}
}
\description{
Return a logical vector indicating which are valid scans (samples).
}
\seealso{
[setValidScan()]
}
