% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getValidScan}
\alias{getValidScan}
\alias{getValidScan,GbsrGenotypeData-method}
\title{Return a logical vector indicating which are valid scans (samples).}
\usage{
getValidScan(object, parents = FALSE, ...)

\S4method{getValidScan}{GbsrGenotypeData}(object, parents = FALSE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{parents}{A logical value to indicate to set FALSE or TRUE to parental samples. If you specify \code{parents = "only"}, this function returns a logical vector indicating TRUE for only parental samples.}

\item{...}{Unused.}
}
\description{
Return a logical vector indicating which are valid scans (samples).
}
\seealso{
\code{\link[=setValidScan]{setValidScan()}}
}
