% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{getValidSnp,GbsrGenotypeData-method}
\alias{getValidSnp,GbsrGenotypeData-method}
\title{Return a logical vector indicating which are valid SNP markers.}
\usage{
\S4method{getValidSnp}{GbsrGenotypeData}(object, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}
}
\description{
Return a logical vector indicating which are valid SNP markers.
}
\seealso{
[setValidSnp()]
}
