% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getValidSnp}
\alias{getValidSnp}
\alias{getValidSnp,GbsrGenotypeData-method}
\title{Return a logical vector indicating which are valid SNP markers.}
\usage{
getValidSnp(object, ...)

\S4method{getValidSnp}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\description{
Return a logical vector indicating which are valid SNP markers.
}
\seealso{
\code{\link[=setValidSnp]{setValidSnp()}}
}
