% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{isOpenGDS}
\alias{isOpenGDS}
\alias{isOpenGDS,GbsrGenotypeData-method}
\title{Check if a GDS file has been opened or not.}
\usage{
isOpenGDS(object, ...)

\S4method{isOpenGDS}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\value{
`TRUE` if the GDS file linked to the input GbsrGenotypeData object has been opened, while `FALSE` if closed.
}
\description{
Check if a GDS file has been opened or not.
}
