% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{loadScanAnnot,GbsrGenotypeData-method}
\alias{loadScanAnnot,GbsrGenotypeData-method}
\title{Load the stored ScanAnnotationDataSet information}
\usage{
\S4method{loadScanAnnot}{GbsrGenotypeData}(object, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object}
}
\description{
All the data stored in the ScanAnnotationDataSet slot of the GbsrGenotypeData
object can be saved in the GDS file linked to the given GbsrGenotypeData object via [saveScanAnnot()].
You can load the saved data using this function.
}
