% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{loadSnpAnnot,GbsrGenotypeData-method}
\alias{loadSnpAnnot,GbsrGenotypeData-method}
\title{Load the stored SnpAnnotationDataSet information}
\usage{
\S4method{loadSnpAnnot}{GbsrGenotypeData}(object, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object}
}
\description{
All the data stored in the SnpAnnotatoinDataSet slot of the GbsrGenotypeData
object can be saved in the GDS file linked to the given GbsrGenotypeData object via [saveSnpAnnot()].
You can load the saved data using this function.
}
