% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{loadSnpAnnot}
\alias{loadSnpAnnot}
\alias{loadSnpAnnot,GbsrGenotypeData-method}
\title{Load the stored \link{SnpAnnotationDataFrame} information}
\usage{
loadSnpAnnot(object, ...)

\S4method{loadSnpAnnot}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object}

\item{...}{Unused.}
}
\value{
A \link{GbsrGenotypeData} object.
}
\description{
All the data stored in the \code{SnpAnnotationnDataFrame} slot
of the \link{GbsrGenotypeData}
object can be saved in the GDS file linked to the given \link{GbsrGenotypeData}
object via \code{\link[=saveSnpAnnot]{saveSnpAnnot()}}. You can load the saved data using this
function.
}
\examples{
# Create a GDS file from a sample VCF file.
vcf_fn <- system.file("extdata", "sample.vcf", package = "GBScleanR")
gds_fn <- tempfile("sample", fileext = ".gds")
gbsrVCF2GDS(vcf_fn = vcf_fn, out_fn = gds_fn, force = TRUE)

# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds <- loadGDS(gds_fn)

# Save data stored in a [ScanAnnotationDataFrame] object
# linked in the slot of the `GbsrGenotypeData.`
gds <- saveSnpAnnot(gds)

# Close the connection to the GDS file
closeGDS(gds)

# Open the connection to the GDS file again.
gds <- loadGDS(gds)

# Load the saved [SnpAnnotationDataFrame] object
gds <- loadSnpAnnot(gds)

# Close the connection to the GDS file
closeGDS(gds)


}
