% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{nscan}
\alias{nscan}
\alias{nscan,GbsrGenotypeData-method}
\title{Return the number of scans (samples).}
\usage{
nscan(object, valid = TRUE, ...)

\S4method{nscan}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\description{
This function returns the number of samples recorded in the GDS file
connected to the given GbsrGenotypeData object.
}
\details{
If `valid = TRUE`, the number of samples which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. You need
the number of over all samples, set `valid = FALSE`.
[getValidSnp()] tells you which samples are valid.
}
\seealso{
[getValidSnp()]
}
