% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{nsnp,GbsrGenotypeData-method}
\alias{nsnp,GbsrGenotypeData-method}
\title{Return the number of SNPs.}
\usage{
\S4method{nsnp}{GbsrGenotypeData}(object, valid = TRUE, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}
}
\description{
This function returns the number of SNPs recorded in the GDS file
connected to the given GbsrGenotypeData object.
}
\details{
If `valid = TRUE`, the number of SNPs which are labeled `TRUE` in
the SnpAnnotationDataSet slot will be returned. You need the number
of over all SNPs, set `valid = FALSE`. [getValidSnp()] tells you
which markers are valid.
}
\seealso{
[getValidSnp()]
}
