% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{nsnp}
\alias{nsnp}
\alias{nsnp,GbsrGenotypeData-method}
\title{Return the number of SNPs.}
\usage{
nsnp(object, valid = TRUE, chr = NULL, ...)

\S4method{nsnp}{GbsrGenotypeData}(object, valid, chr)

\S4method{nsnp}{GbsrGenotypeData}(object, valid, chr)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{valid}{A logical value. See details.}

\item{chr}{A index to spefcify chromosome to get information.}

\item{...}{Unused.}
}
\value{
An integer value to indicate the number of SNP markers.
}
\description{
This function returns the number of SNPs recorded in the GDS file
connected to the given \link{GbsrGenotypeData} object.
}
\details{
If \code{valid = TRUE}, the number of SNPs which are labeled \code{TRUE} in
the \link{SnpAnnotationDataFrame} slot will be returned. You need the number
of over all SNPs, set \code{valid = FALSE}. \code{\link[=getValidSnp]{getValidSnp()}} tells you
which markers are valid.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

nsnp(gds)

# Close the connection to the GDS file.
closeGDS(gds)

}
\seealso{
\code{\link[=getValidSnp]{getValidSnp()}}
}
