% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{pairsGBSR}
\alias{pairsGBSR}
\title{Draw a scatter plot of a pair of specified statistics}
\usage{
pairsGBSR(
  x,
  stats1 = "dp",
  stats2 = "missing",
  target = "snp",
  q = 0.5,
  size = 0.5,
  alpha = 0.8,
  color = c(Marker = "darkblue", Sample = "darkblue"),
  fill = c(Marker = "skyblue", Sample = "skyblue"),
  smooth = FALSE,
  ggargs = NULL
)
}
\arguments{
\item{x}{A \link{GbsrGenotypeData} object.}

\item{stats1}{A string to specify statistics to be drawn.}

\item{stats2}{A string to specify statistics to be drawn.}

\item{target}{Either or both of "snp" and "scan", e.g.
\code{target = "snp"} to draw a histogram only for SNPs.}

\item{q}{An integer to specify a quantile calculated via
\code{\link[=calcReadStats]{calcReadStats()}}.}

\item{size}{A numeric value to specify the dot size of a scatter plot.}

\item{alpha}{A numeric value [0-1] to specify the transparency of
dots in a scatter plot.}

\item{color}{A named vector "Marker" and "Sample" to specify border
color of bins in the histograms.}

\item{fill}{A named vector "Marker" and "Sample" to specify fill color
of bins in the histograms.\verb{stats = "geno} only requires "Ref", "Het"
and "Alt", while others uses the value named "Marker".}

\item{smooth}{A logical value to indicate whether draw a smooth line for
data points. See also \code{\link[ggplot2:geom_smooth]{ggplot2::stat_smooth()}}.}

\item{ggargs}{An expression to be internally passed to a ggplot object
to be drawn.}
}
\value{
NULL.
}
\description{
Draw a scatter plot of a pair of specified statistics
}
\details{
You can draw a scatter plot of per-marker and/or per-sample summary
statistics specified at \code{stats1} and \code{stats2}. The "stats1" and "stats2"
arguments can take the following values:
\itemize{
\item{"missing"}{"Proportion of missing genotype calls."},
\item{"het"}{"Proportion of heterozygote calls."},
\item{"raf"}{"Reference allele frequency."},
\item{"dp"}{"Total read counts."},
\item{"ad_ref"}{"Reference allele read counts."},
\item{"ad_alt"}{"Alternative allele read counts."},
\item{"rrf"}{"Reference allele read frequency."},
\item{"mean_ref"}{"Mean of reference allele read counts."},
\item{"sd_ref"}{"Standard deviation of reference allele read counts."},
\item{"qtile_ref"}{"Quantile of reference allele read counts."},
\item{"mean_alt"}{"Mean of alternative allele read counts."},
\item{"sd_alt"}{"Standard deviation of alternative allele read counts."},
\item{"qtile_alt"}{"Quantile of alternative allele read counts."},
\item{"mq"}{"Mapping quality."},
\item{"fs"}{"Phred-scaled p-value (strand bias)"},
\item{"qd"}{"Variant Quality by Depth"},
\item{"sor"}{"Symmetric Odds Ratio (strand bias)"},
\item{"mqranksum"}{"Alt vs. Ref read mapping qualities"},
\item{"readposranksum"}{"Alt vs. Ref read position bias"},
\item{"baseqranksum"}{"Alt Vs. Ref base qualities"},
}

To draw scatter plots for "missing", "het", "raf", you need to run
\code{\link[=countGenotype]{countGenotype()}} first to obtain statistics. Similary, "dp",
"ad_ref", "ad_alt", "rrf" requires values obtained via \code{\link[=countRead]{countRead()}}.
\code{\link[=calcReadStats]{calcReadStats()}} should be executed before drawing line plots of
"mean_ref", "sd_ref", "qtile_ref", "mean_alt", "sd_alt", and
"qtile_alt". "mq", "fs", "qd", "sor", "mqranksum", "readposranksum",
and "baseqranksum" only work with \code{target = "snp"}, if your data
contains those values supplied via SNP calling tools like
\href{https://gatk.broadinstitute.org/hc/en-us}{GATK}.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)

# Summarize genotype count information to be used in `pairsGBSR()`
gdata <- countGenotype(gdata)

# Draw scatter plots of missing rate vs heterozygosity.
pairsGBSR(gdata, stats1 = "missing", stats2 = "het")

# Close the connection to the GDS file
closeGDS(gdata)


}
