% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotGBSR}
\alias{plotGBSR}
\title{Draw line plots of specified statistics}
\usage{
plotGBSR(
  x,
  stats = c("dp", "missing", "het"),
  coord = NULL,
  q = 0.5,
  lwd = 0.5,
  binwidth = NULL,
  color = c(Marker = "darkblue", Ref = "darkgreen", Het = "magenta", Alt = "blue")
)
}
\arguments{
\item{x}{A \link{GbsrGenotypeData} object.}

\item{stats}{A string to specify statistics to be drawn.}

\item{coord}{A vector with two integer specifying the number of rows
and columns to draw faceted line plots for chromosomes.}

\item{q}{An integer to specify a quantile calculated via
\code{\link[=calcReadStats]{calcReadStats()}}.}

\item{lwd}{A numeric value to specify the line width in plots.}

\item{binwidth}{An integer to specify bin width of the histogram.
This argument only work with \code{stats = "marker"} and is passed to the ggplot
function.}

\item{color}{A strings vector named "Marker", "Ref", "Het", "Alt"
to specify line colors. \verb{stats = "geno} only requires "Ref", "Het" and "Alt",
while others uses the value named "Marker".}
}
\value{
A ggplot object.
}
\description{
Draw line plots of specified statistics
}
\details{
You can draw line plots of several summary statistics of genotype counts
and read counts per sample and per marker. The "stats" argument can take
the following values:
\itemize{
\item{"marker"}{"Marker density."},
\item{"geno"}{"Proportion of missing genotype calls."},
\item{"missing"}{"Proportion of missing genotype calls."},
\item{"het"}{"Proportion of heterozygote calls."},
\item{"raf"}{"Reference allele frequency."},
\item{"dp"}{"Total read counts."},
\item{"ad_ref"}{"Reference allele read counts."},
\item{"ad_alt"}{"Alternative allele read counts."},
\item{"rrf"}{"Reference allele read frequency."},
\item{"mean_ref"}{"Mean of reference allele read counts."},
\item{"sd_ref"}{"Standard deviation of reference allele read counts."},
\item{"qtile_ref"}{"Quantile of reference allele read counts."},
\item{"mean_alt"}{"Mean of alternative allele read counts."},
\item{"sd_alt"}{"Standard deviation of alternative allele read counts."},
\item{"qtile_alt"}{"Quantile of alternative allele read counts."},
\item{"mq"}{"Mapping quality."},
\item{"fs"}{"Phred-scaled p-value (strand bias)"},
\item{"qd"}{"Variant Quality by Depth"},
\item{"sor"}{"Symmetric Odds Ratio (strand bias)"},
\item{"mqranksum"}{"Alt vs. Ref read mapping qualities"},
\item{"readposranksum"}{"Alt vs. Ref read position bias"},
\item{"baseqranksum"}{"Alt Vs. Ref base qualities"},
}

To draw line plots for "missing", "het", "raf", you need to run
\code{\link[=countGenotype]{countGenotype()}} first to obtain statistics. Similary, "dp",
"ad_ref", "ad_alt", "rrf" requires values obtained via \code{\link[=countRead]{countRead()}}.
\code{\link[=calcReadStats]{calcReadStats()}} should be executed before drawing line plots of
"mean_ref", "sd_ref", "qtile_ref", "mean_alt", "sd_alt", and
"qtile_alt". "mq", "fs", "qd", "sor", "mqranksum", "readposranksum",
#' and "baseqranksum" only work with \code{target = "snp"}, if your data
contains those values supplied via SNP calling tools like
\href{https://gatk.broadinstitute.org/hc/en-us}{GATK}.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)

# Summarize genotype count information to be used in `plotGBSR()`
gdata <- countGenotype(gdata)

# Draw line plots of missing rate, heterozygosity, proportion of genotype
# calls per SNP.
plotGBSR(gdata, stats = "missing")

# Calculate means, standard deviations, quantile values of read counts
# to be used in `plotGBSR()`
gdata <- calcReadStats(gdata, q = 0.9)

# Draw line plots of 90 percentile values of reference read counts and
# alternative read counts per SNP and per sample.
plotGBSR(gdata, stats = "qtile_ref", q = 0.9)

# Close the connection to the GDS file
closeGDS(gdata)

}
