% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData_v2.R
\name{plotSmoothedRatio}
\alias{plotSmoothedRatio}
\title{Draw line plots of proportion of reference allele read counts
per marker per sample.}
\usage{
plotSmoothedRatio(x, coord = NULL, chr = NULL, ind = 1, dot_fill = "blue")
}
\arguments{
\item{x}{A \link{GbsrGenotypeData} object.}

\item{coord}{A vector with two integer specifying the number of rows and
columns to draw faceted line plots for chromosomes.}

\item{chr}{A vector of indexes to specify chromosomes to be drawn.}

\item{ind}{An index to specify samples to be drawn.}

\item{dot_fill}{A string to indicate the dot color in a plot.}

\item{node}{Either one of "raw", "filt", and "cor" to output raw
genotype data, filtered genotype data, or corrected genotype data,
respectively.}
}
\value{
A ggplot object.
}
\description{
This function calculate a proportion of reference allele read
counts per marker per sample and
draw line plots of them in facets for each chromosome for each sample.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)

plotReadRatio(gdata, ind = 1)

# Close the connection to the GDS file
closeGDS(gdata)

}
