% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{replaceGDSdata}
\alias{replaceGDSdata}
\alias{replaceGDSdata,GbsrGenotypeData-method}
\title{Replace data stored in the GDS file}
\usage{
replaceGDSdata(object, target, node = "filt", ...)

\S4method{replaceGDSdata}{GbsrGenotypeData}(object, target, node)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{A vector of combinations of "sample.id", "genotype", and "ad".}

\item{node}{Either one of "filt" and "cor" to replace raw genotype data with filtered genotype data or corrected genotype data, respectively.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Replace the original data with the filtered data or replace sample IDs in the GDS file linked
to an input GbsrGenotypeData object.
}
\details{
If \code{target = "genotype"}, replace the data stored in the "genotype" node with the
data in the "filt.genotype" node for genotype call data. If \verb{target = "ad},
The data stored in the "data" node is also replaced with the data in the
"filt.data" node for read count data. \code{target = "sample.id"} makes this function to
replace data in the "sample.id" node with the sample IDs stored in the
ScanAnnotationDataSet slot of the input GbsrGenotypeData object.
}
