% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{resetCallFilters,GbsrGenotypeData-method}
\alias{resetCallFilters,GbsrGenotypeData-method}
\title{Reset the filter made by [setCallFiler()]}
\usage{
\S4method{resetCallFilters}{GbsrGenotypeData}(object, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Return genotype calls and read count data to the original data which are same with
those data before running [setCallFilter()].
}
