% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{resetFilters,GbsrGenotypeData-method}
\alias{resetFilters,GbsrGenotypeData-method}
\title{Reset all filters made by [setScanFiler()], [setSnpFiler()], and [setCallFiler()].}
\usage{
\S4method{resetFilters}{GbsrGenotypeData}(object, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Return all data intact.
}
