% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{resetFilters}
\alias{resetFilters}
\alias{resetFilters,GbsrGenotypeData-method}
\title{Reset all filters made by [setScanFiler()], [setSnpFiler()], and [setCallFiler()].}
\usage{
resetFilters(object, ...)

\S4method{resetFilters}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Return all data intact.
}
