% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{resetScanFilters,GbsrGenotypeData-method}
\alias{resetScanFilters,GbsrGenotypeData-method}
\title{Reset the filter made by [setScanFiler()]}
\usage{
\S4method{resetScanFilters}{GbsrGenotypeData}(object, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Remove "invalid" labels put on samples and make all samples valid.
}
