% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{resetSnpFilters,GbsrGenotypeData-method}
\alias{resetSnpFilters,GbsrGenotypeData-method}
\title{Reset the filter made by [setSnpFiler()]}
\usage{
\S4method{resetSnpFilters}{GbsrGenotypeData}(object, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Remove "invalid" labels put on markers and make all markers valid.
}
