% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{saveScanAnnot}
\alias{saveScanAnnot}
\alias{saveScanAnnot,GbsrGenotypeData-method}
\title{Write out the information stored in the \link{ScanAnnotationDataFrame} slot}
\usage{
saveScanAnnot(object, ...)

\S4method{saveScanAnnot}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object}

\item{...}{Unused.}
}
\value{
NA \link{GbsrGenotypeData} object.
}
\description{
All the data stored in a \link{ScanAnnotationDataFrame} slot of
the \link{GbsrGenotypeData} object can be saved in the GDS file linked
to the given \link{GbsrGenotypeData} object. You can load the saved data
using \code{\link[=loadSnpAnnot]{loadSnpAnnot()}}.
}
\examples{
# Create a GDS file from a sample VCF file.
vcf_fn <- system.file("extdata", "sample.vcf", package = "GBScleanR")
gds_fn <- tempfile("sample", fileext = ".gds")
gbsrVCF2GDS(vcf_fn = vcf_fn, out_fn = gds_fn, force = TRUE)

# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gdata <- loadGDS(gds_fn)

# Save data stored in a [ScanAnnotationDataFrame] object
# linked in the slot of the `GbsrGenotypeData.`
gdata <- saveSnpAnnot(gdata)

# Close the connection to the GDS file
closeGDS(gdata)

}
