% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{setRawGenotype}
\alias{setRawGenotype}
\alias{setRawGenotype,GbsrGenotypeData-method}
\title{Set the origina; data to be used in GBScleanR's functions}
\usage{
setRawGenotype(object, ...)

\S4method{setRawGenotype}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Set the "genotype" node and the "data" node as primary nodes for genotype
data and read count data. The data stored in the primary nodes are used in the
functions of GBScleanR.
}
\details{
A GbsrGenotypeData object storing information of the primary node of genotype data and
read count data. All of the functions implemented in `GBScleanR` check the primary nodes
and use data stored in those nodes. [setCallFilter()] create new nodes storing
filtered genotype calls and read counts in a GDS file and change the primary nodes to
"filt.genotype" and "filt.data" for genotype and read count data, respectively.
`SetRawGenotype()` set back the nodes to the original, those are "genotype" and "data" for
genotype and read count data, respectively. You can set the filtered data again by
[SetFiltGenotype()].
}
