% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{setScanFilter}
\alias{setScanFilter}
\alias{setScanFilter,GbsrGenotypeData-method}
\title{Filter out scans (samples)}
\usage{
setScanFilter(
  object,
  id,
  missing = 1,
  het = c(0, 1),
  mac = 0,
  maf = 0,
  ad_ref = c(0, Inf),
  ad_alt = c(0, Inf),
  dp = c(0, Inf),
  mean_ref = c(0, Inf),
  mean_alt = c(0, Inf),
  sd_ref = Inf,
  sd_alt = Inf,
  ...
)

\S4method{setScanFilter}{GbsrGenotypeData}(
  object,
  id,
  missing = 1,
  het = c(0, 1),
  mac = 0,
  maf = 0,
  ad_ref = c(0, Inf),
  ad_alt = c(0, Inf),
  dp = c(0, Inf),
  mean_ref = c(0, Inf),
  mean_alt = c(0, Inf),
  sd_ref = Inf,
  sd_alt = Inf
)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{id}{A vector of strings match with scan ID which can be retrieve by `getScanID()`.}

\item{missing}{A numeric value [0-1] to specify the maximum missing genotype call rate per sample.}

\item{het}{A numeric value [0-1] to specify the maximum heterozygous genotype call rate per sample.}

\item{mac}{A integer value to specify the minimum minor allele count per sample.}

\item{maf}{A numeric value to specify the minimum minor allele frequency per sample.}

\item{ad_ref}{A numeric vector with length two specifying lower and upper limit of reference read counts per sample.}

\item{ad_alt}{A numeric vector with length two specifying lower and upper limit of alternative read counts per sample.}

\item{dp}{A numeric vector with length two specifying lower and upper limit of total read counts per sample.}

\item{mean_ref}{A numeric vector with length two specifying lower and upper limit of mean of reference read counts per sample.}

\item{mean_alt}{A numeric vector with length two specifying lower and upper limit of mean of alternative read counts per sample.}

\item{sd_ref}{A numeric value specifying the upper limit of standard deviation of reference read counts per sample.}

\item{sd_alt}{A numeric value specifying the upper limit of standard deviation of alternative read counts per sample.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Search samples which do not meet the criteria and label them as "invalid".
}
\details{
For `mean_ref`, `mean_alt`, `sd_ref`, and `sd_alt`, this function calculate mean and
standard deviation of reads obtained at SNP markers of each sample. If a mean read counts
of a sample was smaller than the specified lower limit or larger than the upper limit,
this function labels the sample as "invalid". In the case of `sd_ref` and `sd_alt`,
standard deviations of read counts of each sample are checked and the samples having a
larger standard deviation will be labeled as "invalid". To check valid and invalid
samples, run [getValidScan()].
}
\examples{
\dontrun{
gds <- loadGDS("/path/to/GDS.gds")
gds <- setScanFilter(gds, id = getScanID(gds)[1:10], missing = 0.2, dp = c(5, Inf))
}

}
