% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData_v2.R
\name{smoothReadRatio}
\alias{smoothReadRatio}
\alias{smoothReadRatio,GbsrGenotypeData-method}
\title{Calculate smoothed read ratio per marker.}
\usage{
smoothReadRatio(
  object,
  node = "raw",
  window = 1e+05,
  n_threads = 1,
  verbose = TRUE,
  ...
)

\S4method{smoothReadRatio}{GbsrGenotypeData}(object, node, window, n_threads, verbose)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{window}{A numeric to indicate the size of window to calculate
smoothed read ratio.}

\item{n_threads}{An integer value to specify the number of
threads used for the calculation. The default is 1 and if \code{n_threads = NULL},
automatically set half the number of available threads on the computer.}

\item{...}{Unused.}
}
\value{
A \link{GbsrGenotypeData} object.
}
\description{
Calculate smoothed read ratio per marker.
}
\details{
Calculate smoothed read ratio per marker.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

gds <- smoothReadRatio(gds)

# Close the connection to the GDS file.
closeGDS(gds)

}
