% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-GbsrGenotypeData.R
\name{thinMarker,GbsrGenotypeData-method}
\alias{thinMarker,GbsrGenotypeData-method}
\title{Remove markers potentially having redundant information.}
\usage{
\S4method{thinMarker}{GbsrGenotypeData}(object, range = 150, ...)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{range}{A integer value to indicate the stretch to search markers.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Markers within the length of the sequenced reads (usually ~ 150 bp, up to your sequencer)
potentially have redundant information and those will cause unexpected errors
in error correction which assumes independency of markers each other.
This function only retains the first marker or the least missing rate marker
from the markers locating within the specified stretch.
}
\details{
This function search valid markers from the first marker of each chromosome and
compare its physical position with a neighbor marker. If the distance between those
markers are equal or less then `range`, one of them which has a larger missing rate
will be removed (labeled as invalid marker). When the first marker was retained and
the second marker was removed as invalid marker, next the distance between the first marker
and the third marker will be checked and this cycle is repeated until reaching the
end of each chromosome. Run [getValidSnp()] to check the valid SNP markers.
}
\examples{
gds <- loadGDS("/path/to/GDS.gds")
gds <- thinMarker(gds, range = 150)

}
