% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{addScan}
\alias{addScan}
\alias{addScan,GbsrGenotypeData-method}
\title{Add genotype data into the GDS file}
\usage{
addScan(object, id, genotype, reads, ...)

\S4method{addScan}{GbsrGenotypeData}(object, id, genotype, reads)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{id}{A character vector.}

\item{genotype}{A numeric vector or matrix indicating genotypes at markers
of given samples. The length or the number of columns should match with the
number of markers recorded in the GDS file, which can be obtained via
nsnp() with \code{valid = FALSE}.}

\item{reads}{A numeric vector of matrix indicating read counts at markers of
given samples. The length or the number of columns should match with twice
the number of markers recorded in the GDS file, which can be obtained via
nsnp() with \code{valid = FALSE}.}

\item{...}{Unused.}
}
\value{
A \link{GbsrGenotypeData} object.
}
\description{
Add genotype data into the GDS file
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

id <- "Dummy1"
genotype <- sample(c(0:3), nsnp(gds), replace = TRUE)
reads <- round(rexp(nsnp(gds) * 2, 1/5))
missing_pos <- which(genotype == 3)
reads[c(missing_pos, missing_pos + 1)] <- 0
ref_pos <- which(genotype == 2)
reads[ref_pos + 1] <- 0
alt_pos <- which(genotype == 0)
reads[alt_pos] <- 0

gds <- addScan(gds, id, genotype, reads)

# Close the connection to the GDS file
closeGDS(gds)

}
