% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{boxplotGBSR}
\alias{boxplotGBSR}
\title{Draw boxplots of specified statistics}
\usage{
boxplotGBSR(
  x,
  stats = "missing",
  target = c("snp", "scan"),
  q = 0.5,
  color = c(Marker = "darkblue", Sample = "darkblue"),
  fill = c(Marker = "skyblue", Sample = "skyblue"),
  ggargs = NULL
)
}
\arguments{
\item{x}{A GbsrGenotypeData object.}

\item{stats}{A vector of strings to specify statistics to be drawn.}

\item{target}{Either or both of "snp" and "scan", e.g. `target = "snp"` to draw a histogram only for SNPs.}

\item{q}{An integer to specify a quantile calculated via [calcReadStats()].}

\item{color}{A named vector "Marker" and "Sample" to specify border color of bins in the histograms.}

\item{fill}{A named vector "Marker" and "Sample" to specify fill color of bins in the histograms.}

\item{ggargs}{An expression to be internally passed to a ggplot object to be drawn.}
}
\description{
Draw boxplots of specified statistics
}
\details{
You can draw boxplots of several summary statistics of genotype counts
and read counts per sample and per marker. The "stats" argument can take
the following values:

\itemize{
\item{"missing"}{"Proportion of missing genotype calls."},
\item{"het"}{"Proportion of heterozygote calls."},
\item{"raf"}{"Reference allele frequency."},
\item{"dp"}{"Total read counts."},
\item{"ad_ref"}{"Reference allele read counts."},
\item{"ad_alt"}{"Alternative allele read counts."},
\item{"rrf"}{"Reference allele read frequency."},
\item{"mean_ref"}{"Mean of reference allele read counts."},
\item{"sd_ref"}{"Standard deviation of reference allele read counts."},
\item{"qtile_ref"}{"Quantile of reference allele read counts."},
\item{"mean_alt"}{"Mean of alternative allele read counts."},
\item{"sd_alt"}{"Standard deviation of alternative allele read counts."},
\item{"qtile_alt"}{"Quantile of alternative allele read counts."},
\item{"mq"}{"Mapping quality."},
\item{"fs"}{"Phred-scaled p-value (strand bias)"},
\item{"qd"}{"Variant Quality by Depth"},
\item{"sor"}{"Symmetric Odds Ratio (strand bias)"},
\item{"mqranksum"}{"Alt vs. Ref read mapping qualities"},
\item{"readposranksum"}{"Alt vs. Ref read position bias"},
\item{"baseqranksum"}{"Alt Vs. Ref base qualities"},
}

To draw boxplots for "missing", "het", "raf", you need to run [countGenotype()]
first to obtain statistics. Similary, "dp", "ad_ref", "ad_alt", "rrf" requires
values obtained via [countRead()]. [calcReadStats()] should be executed before
drawing boxplots of "mean_ref", "sd_ref", "qtile_ref", "mean_alt", "sd_alt",
and "qtile_alt". "mq", "fs", "qd", "sor", "mqranksum", "readposranksum",
and "baseqranksum" only work with `target = "snp"`, if your data contains those
values supplied via SNP calling tools like [GATK](https://gatk.broadinstitute.org/hc/en-us).
}
\examples{
\dontrun{
# Draw boxplots of missing rate, heterozygosity, and reference 
# allele frequency per SNP and per sample.
gdata <- loadGDS("/path/to/GDS.gds")
gdata <- countGenotype(gdata)
boxplot(gdata, stats = c("missing", "het", "raf"))

# Draw boxplots of 90 percentile values of reference read counts and
# alternative read counts per SNP and per sample.
gdata <- calcReadStats(gdata, q = 0.9)
boxplot(gdata, stats = c("qtile_ref", "qtile_alt"), q = 0.9)
}

}
