% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{boxplotGBSR}
\alias{boxplotGBSR}
\title{Draw boxplots of specified statistics}
\usage{
boxplotGBSR(
  x,
  stats = "missing",
  target = c("snp", "scan"),
  q = 0.5,
  color = c(Marker = "darkblue", Sample = "darkblue"),
  fill = c(Marker = "skyblue", Sample = "skyblue")
)
}
\arguments{
\item{x}{A \link{GbsrGenotypeData} object.}

\item{stats}{A string to specify statistics to be drawn.}

\item{target}{Either or both of "snp" and "scan", e.g. \code{target = "snp"}
to draw a histogram only for SNPs.}

\item{q}{An integer to specify a quantile calculated via
\code{\link[=calcReadStats]{calcReadStats()}}.}

\item{color}{A named vector "Marker" and "Sample" to specify
border color of bins in the histograms.}

\item{fill}{A named vector "Marker" and "Sample" to specify
fill color of bins in the histograms.}
}
\value{
A ggplot object.
}
\description{
Draw boxplots of specified statistics
}
\details{
You can draw boxplots of several summary statistics of genotype counts
and read counts per sample and per marker. The "stats" argument can take
the following values:

\itemize{
\item{"missing"}{"Proportion of missing genotype calls."},
\item{"het"}{"Proportion of heterozygote calls."},
\item{"raf"}{"Reference allele frequency."},
\item{"dp"}{"Total read counts."},
\item{"ad_ref"}{"Reference allele read counts."},
\item{"ad_alt"}{"Alternative allele read counts."},
\item{"rrf"}{"Reference allele read frequency."},
\item{"mean_ref"}{"Mean of reference allele read counts."},
\item{"sd_ref"}{"Standard deviation of reference allele read counts."},
\item{"qtile_ref"}{"Quantile of reference allele read counts."},
\item{"mean_alt"}{"Mean of alternative allele read counts."},
\item{"sd_alt"}{"Standard deviation of alternative allele read counts."},
\item{"qtile_alt"}{"Quantile of alternative allele read counts."},
\item{"mq"}{"Mapping quality."},
\item{"fs"}{"Phred-scaled p-value (strand bias)"},
\item{"qd"}{"Variant Quality by Depth"},
\item{"sor"}{"Symmetric Odds Ratio (strand bias)"},
\item{"mqranksum"}{"Alt vs. Ref read mapping qualities"},
\item{"readposranksum"}{"Alt vs. Ref read position bias"},
\item{"baseqranksum"}{"Alt Vs. Ref base qualities"},
}

To draw boxplots for "missing", "het", "raf", you need to run
\code{\link[=countGenotype]{countGenotype()}} first to obtain statistics. Similary, "dp",
"ad_ref", "ad_alt", "rrf" requires values obtained via \code{\link[=countRead]{countRead()}}.
\code{\link[=calcReadStats]{calcReadStats()}} should be executed before drawing boxplots of
"mean_ref", "sd_ref", "qtile_ref", "mean_alt", "sd_alt", and
"qtile_alt". "mq", "fs", "qd", "sor", "mqranksum", "readposranksum",
and "baseqranksum" only work with \code{target = "snp"}, if your data
contains those values supplied via SNP calling tools like
\href{https://gatk.broadinstitute.org/hc/en-us}{GATK}.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)

# Summarize genotype count information to be used in `boxplotGBSR()`
gdata <- countGenotype(gdata)

boxplotGBSR(gdata, stats = "missing")

# Calculate means, standard deviations, quantile values of read counts
# to be used in `boxplotGBSR()`
gdata <- calcReadStats(gdata, q = 0.9)

# Draw boxplots of 90 percentile values of reference read counts and
# alternative read counts per SNP and per sample.
boxplotGBSR(gdata, stats = "qtile_ref", q = 0.9)

# Close the connection to the GDS file
closeGDS(gdata)

}
