% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{closeGDS}
\alias{closeGDS}
\alias{closeGDS,GbsrGenotypeData-method}
\title{Close the connection to the GDS file}
\usage{
closeGDS(object, ...)

\S4method{closeGDS}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{...}{Unused.}
}
\value{
NULL.
}
\description{
Close the connection to the GDS file linked to the given
\link{GbsrGenotypeData} object.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

# Close the connection to the GDS file
closeGDS(gds)

}
