% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{gbsrGDS2VCF}
\alias{gbsrGDS2VCF}
\alias{gbsrGDS2VCF,GbsrGenotypeData-method}
\title{Write a VCF file based on data in a GDS file}
\usage{
gbsrGDS2VCF(object, out_fn, node = "raw", incl_parents = TRUE, ...)

\S4method{gbsrGDS2VCF}{GbsrGenotypeData}(object, out_fn, node, incl_parents)

\S4method{gbsrGDS2VCF}{GbsrGenotypeData}(object, out_fn, node, incl_parents)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{out_fn}{A string to specify the path to an output VCF file.}

\item{node}{Either one of "raw", "filt", and "cor" to output raw
genotype data, filtered genotype data, or corrected genotype data,
respectively.}

\item{incl_parents}{A logical value to specify whether parental
samples should be included in an output VCF file or not.}

\item{...}{Unused.}
}
\value{
The path to the VCF file.
}
\description{
Write out a VCF file with raw, filtered, or corrected genotype data
stored in a GDS file. The output VCF file contains only the GT filed,
while other annotations, AD, DP and other information will be omitted.
}
\details{
Create a VCF file at location specified by out_fn.
The connection to the GDS file of the input \link{GbsrGenotypeData} object will be
automatically closed for internal file handling in this function. Please use
\code{\link[=openGDS]{openGDS()}} to open the connection again. If you use \code{\link[=loadGDS]{loadGDS()}}, summary
statistics and filtering information will be discarded.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

# Create a VCF file with data from the GDS file
#  connected to the [GbsrGenotypeData] oobject.
out_fn <- tempfile("sample_out", fileext = ".vcf.gz")
gbsrGDS2VCF(gds, out_fn)

}
