% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUitlities.R
\name{gbsrVCF2GDS}
\alias{gbsrVCF2GDS}
\title{Convert a VCF file to a GDS file}
\usage{
gbsrVCF2GDS(vcf_fn, out_fn, force = FALSE)
}
\arguments{
\item{vcf_fn}{A string to indicate path to an input VCF file.}

\item{out_fn}{A string to indicate path to an output GDS file.}

\item{force}{A logical value to overwrite a GDS file even if the file specified in "out_fn" exists.}
}
\value{
The output GDS file path.
}
\description{
This function converts a variant call data in the VCF format.
The current implementation only accepts biallelic single nucleotide polymorphisms.
Please filter out variants which are insertions and deletions or multiallelic.
You may use "bcftools" or "vcftools" for filtering.
}
\details{
gbsrVCF2GDS converts a VCF file to a GDS file.
The data structure of the GDS file created via this functions is same with
those created by snpgdsVCF2GDS of
\href{https://www.bioconductor.org/packages/release/bioc/html/SNPRelate.html}{SNPRelate}.
Unlike the GDS files created by SNPRelate GBScleanR's GDS files: \cr
\itemize{
\item{"Include annotation information of variants recorded in the INFO filed of the input VCF."}
\item{"Include allelic read count data (indicated as AD) recoreded in the FORMAT filed of the input VCF."}
}
}
