% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getAlleleB}
\alias{getAlleleB}
\alias{getAlleleB,GbsrGenotypeData-method}
\title{Obtain alternative allele information of each SNP marker}
\usage{
getAlleleB(object, valid = TRUE, chr = NULL, ...)

\S4method{getAlleleB}{GbsrGenotypeData}(object, valid, chr)

\S4method{getAlleleB}{GbsrGenotypeData}(object, valid, chr)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{valid}{A logical value. See details.}

\item{chr}{A index to spefcify chromosome to get information.}

\item{...}{Unused.}
}
\value{
A character vector indicating the alternative alleles.
}
\description{
This function returns alternative alleles,
either of A, T, G, and C, of SNP markers.
}
\details{
If \code{valid = TRUE}, the chromosome information of markers which are
labeled \code{TRUE} in the \link{ScanAnnotationDataFrame} slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

getAlleleB(gds)

# Close the connection to the GDS file.
closeGDS(gds)

}
