% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getCountReadAlt}
\alias{getCountReadAlt}
\alias{getCountReadAlt,GbsrGenotypeData-method}
\title{Obtain total alternative read counts per SNP or per scan (sample)}
\usage{
getCountReadAlt(object, target = "snp", valid = TRUE, prop = FALSE, ...)

\S4method{getCountReadAlt}{GbsrGenotypeData}(object, target = "snp", valid = TRUE, prop = FALSE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{valid}{A logical value. See details.}

\item{prop}{A logical value whether to return values as proportions of total alternative read counts in total read counts per SNP or not.}

\item{...}{Unused.}
}
\description{
Obtain total alternative read counts per SNP or per scan (sample)
}
\details{
You need to execute [countRead()] to calculate sumaary statisticsto be
obtained via this function.
If `valid = TRUE`, the chromosome information of markers which are labeled `TRUE`
in the ScanAnnotationDataSet slot will be returned. [getValidSnp()] tells you
which samples are valid.
}
