% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getHaplotype}
\alias{getHaplotype}
\alias{getHaplotype,GbsrGenotypeData-method}
\title{Get haplotype call data.}
\usage{
getHaplotype(object, chr = NULL, parents = FALSE, ...)

\S4method{getHaplotype}{GbsrGenotypeData}(object, chr, parents)
}
\arguments{
\item{object}{A gbsrGenotypeData object.}

\item{chr}{A integer vector of indexes indicating chromosomes to get read count data.}

\item{parents}{A logical value or "only" to include data for parents or to get data only for parents.}

\item{...}{Unused.}
}
\description{
Haplotype calls are retrieved from the GDS file linked to the given
GbsrGenotypeData object.
}
\details{
Haplotype call data can be obtained from the "estimated.haplotype" node of
the GDS file which can be generated via the [clean()] function. Thus, this function
is valid only after having executed [clean()].
}
\seealso{
[clean()]
}
