% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getMeanReadRef}
\alias{getMeanReadRef}
\alias{getMeanReadRef,GbsrGenotypeData-method}
\title{Obtain mean values of total reference read counts per
SNP or per scan (sample)}
\usage{
getMeanReadRef(object, target = "snp", valid = TRUE, ...)

\S4method{getMeanReadRef}{GbsrGenotypeData}(object, target = "snp", valid = TRUE)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{target}{Either of "snp" and "scan".}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A numeric vector of the mean values of
reference allele reads per marker.
}
\description{
Obtain mean values of total reference read counts per
SNP or per scan (sample)
}
\details{
You need to execute \code{\link[=calcReadStats]{calcReadStats()}} to calculate sumaary statisticsto be
obtained via this function.
If \code{valid = TRUE}, the chromosome information of markers which are
labeled \code{TRUE} in the \link{ScanAnnotationDataFrame} slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
gdata <- calcReadStats(gdata)
getMeanReadRef(gdata)
closeGDS(gdata) # Close the connection to the GDS file

}
