% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getMedianReadAlt}
\alias{getMedianReadAlt}
\alias{getMedianReadAlt,GbsrGenotypeData-method}
\title{Obtain quantile values of total alternative read counts per
SNP or per scan (sample)}
\usage{
getMedianReadAlt(object, target = "marker", valid = TRUE, ...)

\S4method{getMedianReadAlt}{GbsrGenotypeData}(object, target, q, valid)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{target}{Either of "snp" and "scan".}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A numeric vector of the quantile values of
alternative allele reads per marker.
}
\description{
Obtain quantile values of total alternative read counts per
SNP or per scan (sample)
}
\details{
You need to execute \code{\link[=calcReadStats]{calcReadStats()}} to calculate sumaary statisticsto be
obtained via this function.
If \code{valid = TRUE}, the chromosome information of markers which are
labeled \code{TRUE} in the \link{ScanAnnotationDataFrame} slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

# Calculate means, standard deviations, quantiles of read counts
# per marker and per sample with or without standardization of
# the counts and store them in the
# [SnpAnnotationDataFrame] and [ScanAnnotationDataFrame] objects
# linked at the slots of the [GbsrGenotypeData] object.
gds <- calcReadStats(gds)

getMedianReadAlt(gds)

# Close the connection to the GDS file.
closeGDS(gds)

}
