% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getParents}
\alias{getParents}
\alias{getParents,GbsrGenotypeData-method}
\title{Get parental sample information}
\usage{
getParents(object, ...)

\S4method{getParents}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{...}{Unused.}
}
\value{
A data frame of parents information indicating scanIDs, memberIDs
and indexes of parental lines assigned via \code{\link[=setParents]{setParents()}}.
}
\description{
This function returns scan IDs, member IDs and indexes of parental samples
set via \code{\link[=setParents]{setParents()}}. Scan IDs are IDs given by user or obtained from the
original VCF file. Member IDs are serial numbers assigned by \code{\link[=setParents]{setParents()}}.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)

# Find the IDs of parental samples.
parents <- grep("Founder", getScanID(gdata), value = TRUE)

# Set the parents.
gdata <- setParents(gdata, parents = parents, flip = TRUE)

# Get the information of parents.
getParents(gdata)

# Close the connection to the GDS file.
closeGDS(gdata)


}
