% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getParents}
\alias{getParents}
\alias{getParents,GbsrGenotypeData-method}
\title{Get parental sample information}
\usage{
getParents(object, ...)

\S4method{getParents}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\description{
This function returns scan IDs, member IDs and indexes of parental samples
set via [setParents()]. Scan IDs are IDs given by user or obtained from the
original VCF file. Member IDs are serial numbers assigned by [setParents()].
}
\examples{
\dontrun{
gds <- loadGDS("/path/to/GDS.gds")
gds <- setParents(gds, parents = c("parent1", "parent2"))
getParents(gds)
}


}
