% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getParents}
\alias{getParents}
\alias{getParents,GbsrGenotypeData-method}
\title{Get parental sample information}
\usage{
getParents(object, ...)

\S4method{getParents}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\description{
This function returns scan IDs, member IDs and indexes of parental samples
set via \code{\link[=setParents]{setParents()}}. Scan IDs are IDs given by user or obtained from the
original VCF file. Member IDs are serial numbers assigned by \code{\link[=setParents]{setParents()}}.
}
\examples{
gds_fn <- system.file("extdata", "simpop.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)
parents <- grep("Founder", getScanID(gds), value = TRUE)
gds <- setParents(gds, parents = parents)
getParents(gds)


}
