% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getPloidy}
\alias{getPloidy}
\alias{getPloidy,GbsrGenotypeData-method}
\title{Get ploidy}
\usage{
getPloidy(object, ...)

\S4method{getPloidy}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}
}
\value{
An integer value.
}
\description{
Get the ploidy of a given population in the input \link{GbsrGenotypeData} object.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

ploidy <- getPloidy(gds)
print(ploidy)

# Close the connection to the GDS file.
closeGDS(gds)

}
\seealso{
\code{\link[=setPloidy]{setPloidy()}}
}
