% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getPosition}
\alias{getPosition}
\alias{getPosition,GbsrGenotypeData-method}
\title{Obtain physical position information of each SNP marker}
\usage{
getPosition(object, valid = TRUE, ...)

\S4method{getPosition}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A integer vetor indicating the physical positions of markers.
}
\description{
This function returns physical positions of SNP markers.
}
\details{
If \code{valid = TRUE}, the chromosome information of markers which
are labeled \code{TRUE} in the \link{ScanAnnotationDataFrame} slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)

getPosition(gdata)

# Close the connection to the GDS file.
closeGDS(gdata)

}
