% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getRead}
\alias{getRead}
\alias{getRead,GbsrGenotypeData-method}
\title{Get read count data.}
\usage{
getRead(object, valid = TRUE, chr = NULL, node = "raw", parents = FALSE, ...)

\S4method{getRead}{GbsrGenotypeData}(object, valid, chr, node, parents)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{valid}{A logical value. See details.}

\item{chr}{A integer vector of indexes indicating
chromosomes to get read count data.}

\item{node}{Either of "raw" and "filt". See details.}

\item{parents}{A logical value or "only" to include data
for parents or to get data only for parents.}

\item{...}{Unused.}
}
\value{
A named list with two elements "ref" and "alt" storing
a matrix of reference allele read counts and a matrix of
alternative read counts for all markers in all samples.
}
\description{
Read counts for reference allele and alternative allele are retrieved
from the GDS file linked to the given \link{GbsrGenotypeData} object.
}
\details{
Read count data can be obtained from the "annotation/format/AD/data" node,
the "annotation/format/AD/norm" node or the
"annotation/format/AD/filt.data" node of the GDS file with
\code{node = "raw"},  \code{node = "norm"}, or \code{node = "filt"}, respectively.
The \code{\link[=calcReadStats]{calcReadStats()}} function generate normalized read count data
in the "annotation/format/AD/norm" node, while \code{\link[=setCallFilter]{setCallFilter()}}
function generate filtered read count data
in the "annotation/format/AD/filt.data" node which can be accessed as
mentioned above.
If \code{valid = TRUE}, read counts for only valid marker and valid samples will
be obtained.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

read <- getRead(gds)

# Close the connection to the GDS file.
closeGDS(gds)

}
\seealso{
\code{\link[=setCallFilter]{setCallFilter()}}
}
