% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getScanID}
\alias{getScanID}
\alias{getScanID,GbsrGenotypeData-method}
\title{Obtain scan (sample) ID}
\usage{
getScanID(object, valid = TRUE, ...)

\S4method{getScanID}{GbsrGenotypeData}(object, valid)

\S4method{getScanID}{GbsrGenotypeData}(object, valid)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A character vector of scan(sample) IDs.
}
\description{
This function returns scan (sample) ID.
}
\details{
If \code{valid = TRUE}, the chromosome information of markers which are
labeled \code{TRUE} in the \link{ScanAnnotationDataFrame} slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

getScanID(gds)

# Close the connection to the GDS file.
closeGDS(gds)

}
