% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getScanID}
\alias{getScanID}
\alias{getScanID,GbsrGenotypeData-method}
\title{Obtain scan (sample) ID}
\usage{
getScanID(object, valid = TRUE, ...)

\S4method{getScanID}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A character vector of scan(sample) IDs.
}
\description{
This function returns scan (sample) ID.
}
\details{
If \code{valid = TRUE}, the chromosome information of markers which are
labeled \code{TRUE} in the ScanAnnotationDataSet slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
gds_fn <- system.file("extdata", "simpop.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
getScanID(gdata)

}
