% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getSDReadRef}
\alias{getSDReadRef}
\alias{getSDReadRef,GbsrGenotypeData-method}
\title{Obtain standard deviations of total reference read counts
per SNP or per scan (sample)}
\usage{
getSDReadRef(object, target = "snp", valid = TRUE, ...)

\S4method{getSDReadRef}{GbsrGenotypeData}(object, target = "snp", valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{target}{Either of "snp" and "scan".}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
A numeric vector of the standard deviations of
reference allele reads per marker.
}
\description{
Obtain standard deviations of total reference read counts
per SNP or per scan (sample)
}
\details{
You need to execute \code{\link[=calcReadStats]{calcReadStats()}} to calculate summary statistics to be
obtained via this function.
If \code{valid = TRUE}, the chromosome information of markers which are
labeled \code{TRUE} in the ScanAnnotationDataSet slot will be returned.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
gdata <- calcReadStats(gdata)
getSDReadRef(gdata)
closeGDS(gdata) # Close the connection to the GDS file

}
