% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getValidScan}
\alias{getValidScan}
\alias{getValidScan,GbsrGenotypeData-method}
\title{Return a logical vector indicating which are valid scans (samples).}
\usage{
getValidScan(object, parents = FALSE, ...)

\S4method{getValidScan}{GbsrGenotypeData}(object, parents)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{parents}{A logical value to indicate to set FALSE or TRUE
to parental samples. If you specify \code{parents = "only"}, this
function returns a logical vector indicating TRUE for only parental samples.}

\item{...}{Unused.}
}
\value{
A logical vector of the same length
with the number of total samples.
}
\description{
Return a logical vector indicating which are valid scans (samples).
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

getValidScan(gds)

# Close the connection the GDS file.
closeGDS(gds)

}
\seealso{
\code{\link[=setValidScan]{setValidScan()}}
}
