% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getValidSnp}
\alias{getValidSnp}
\alias{getValidSnp,GbsrGenotypeData-method}
\title{Return a logical vector indicating which are valid SNP markers.}
\usage{
getValidSnp(object, ...)

\S4method{getValidSnp}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\value{
A logical vector of the same length with
the number of total SNP markers
}
\description{
Return a logical vector indicating which are valid SNP markers.
}
\examples{
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
getValidSnp(gdata)
closeGDS(gdata) # Close the connection to the GDS file

}
\seealso{
\code{\link[=setValidSnp]{setValidSnp()}}
}
