% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{getValidSnp}
\alias{getValidSnp}
\alias{getValidSnp,GbsrGenotypeData-method}
\title{Return a logical vector indicating which are valid SNP markers.}
\usage{
getValidSnp(object, chr = NULL, ...)

\S4method{getValidSnp}{GbsrGenotypeData}(object, chr)

\S4method{getValidSnp}{GbsrGenotypeData}(object, chr)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{chr}{A index to spefcify chromosome to get information.}

\item{...}{Unused.}
}
\value{
A logical vector of the same length with
the number of total SNP markers
}
\description{
Return a logical vector indicating which are valid SNP markers.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

getValidSnp(gds)

# Close the connection to the GDS file.
closeGDS(gds)

}
\seealso{
\code{\link[=setValidSnp]{setValidSnp()}}
}
