% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUitlities.R
\name{loadGDS}
\alias{loadGDS}
\title{Load a GDS file and construct a GbsrGenotypeData object.}
\usage{
loadGDS(gds_fn, non_autosomes = NULL, genotypeData)
}
\arguments{
\item{gds_fn}{A string of the path to an input GDS file.}

\item{non_autosomes}{A list with elements X, Y, XY, and/or M. See details.}

\item{genotypeData}{A GbsrGenotypeData object to reload.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Load data stored in an input GDS file to R environment and
create a GbsrGenotypeData instance.
GBScleanR handles only one class GbsrGenotypeData and
conducts all data manipulation via class methods for it.
}
\examples{
gds_fn <- gbsrVCF2GDS(vcf_fn = "/path/to/vcf.vcf", out_fn = "/path/to/gds.gds/", force = TRUE)
gdata <- loadGDS(gds_fn = gds_fn, non_autosomes = NULL)

# If you would like to reload a GDS file.
gdata <- loadGDS(genotypeData = gdata)

}
