% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtilities.R
\name{loadGDS}
\alias{loadGDS}
\title{Load a GDS file and construct a \code{GbsrGenotypeData} object.}
\usage{
loadGDS(x, verbose = TRUE)
}
\arguments{
\item{x}{A string of the path to an input GDS file or
a \code{GbsrGenotypeData} object to reload.}

\item{verbose}{if TRUE, show information.}
}
\value{
A \code{GbsrGenotypeData} object.
}
\description{
Load data stored in an input GDS file to R environment and
create a \code{GbsrGenotypeData} instance.
GBScleanR handles only one class \code{GbsrGenotypeData} and
conducts all data manipulation via class methods for it.
}
\examples{
# Create a GDS file from a sample VCF file.
vcf_fn <- system.file("extdata", "sample.vcf", package = "GBScleanR")
gds_fn <- tempfile("sample", fileext = ".gds")
gbsrVCF2GDS(vcf_fn = vcf_fn, out_fn = gds_fn, force = TRUE)

# Load data in the GDS file and instantiate a `GbsrGenotypeData` object.
gdata <- loadGDS(gds_fn)

# Reload data from the GDS file.
gdata <- loadGDS(gdata)

# Close the connection to the GDS file.
closeGDS(gdata)

}
