% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{loadScanAnnot}
\alias{loadScanAnnot}
\alias{loadScanAnnot,GbsrGenotypeData-method}
\title{Load the stored ScanAnnotationDataSet information}
\usage{
loadScanAnnot(object, ...)

\S4method{loadScanAnnot}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object}

\item{...}{Unused.}
}
\description{
All the data stored in the ScanAnnotationDataSet slot of the GbsrGenotypeData
object can be saved in the GDS file linked to the given GbsrGenotypeData object via \code{\link[=saveScanAnnot]{saveScanAnnot()}}.
You can load the saved data using this function.
}
