% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{loadSnpAnnot}
\alias{loadSnpAnnot}
\alias{loadSnpAnnot,GbsrGenotypeData-method}
\title{Load the stored SnpAnnotationDataSet information}
\usage{
loadSnpAnnot(object, ...)

\S4method{loadSnpAnnot}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object.
}
\description{
All the data stored in the SnpAnnotatoinDataSet slot of the GbsrGenotypeData
object can be saved in the GDS file linked to the given GbsrGenotypeData
object via \code{\link[=saveSnpAnnot]{saveSnpAnnot()}}. You can load the saved data using this
function.
}
\examples{
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
gdata <- loadSnpAnnot(gdata)
closeGDS(gdata) # Close the connection to the GDS file


}
