% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{nscan}
\alias{nscan}
\alias{nscan,GbsrGenotypeData-method}
\title{Return the number of scans (samples).}
\usage{
nscan(object, valid = TRUE, ...)

\S4method{nscan}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
An integer value to indicate the number of samples.
}
\description{
This function returns the number of samples recorded in the GDS file
connected to the given GbsrGenotypeData object.
}
\details{
If \code{valid = TRUE}, the number of samples which are labeled \code{TRUE}
in the ScanAnnotationDataSet slot will be returned. You need
the number of over all samples, set \code{valid = FALSE}.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
nscan(gdata)
closeGDS(gdata) # Close the connection to the GDS file

}
\seealso{
\code{\link[=getValidScan]{getValidScan()}}
}
