% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{nscan}
\alias{nscan}
\alias{nscan,GbsrGenotypeData-method}
\title{Return the number of scans (samples).}
\usage{
nscan(object, valid = TRUE, ...)

\S4method{nscan}{GbsrGenotypeData}(object, valid)

\S4method{nscan}{GbsrGenotypeData}(object, valid)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
An integer value to indicate the number of samples.
}
\description{
This function returns the number of samples recorded in the GDS file
connected to the given \link{GbsrGenotypeData} object.
}
\details{
If \code{valid = TRUE}, the number of samples which are labeled \code{TRUE}
in the \link{ScanAnnotationDataFrame} slot will be returned. You need
the number of over all samples, set \code{valid = FALSE}.
\code{\link[=getValidSnp]{getValidSnp()}} tells you which samples are valid.
}
\examples{
# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gds <- loadGDS(gds_fn)

nscan(gds)

# Close the connection to the GDS file.
closeGDS(gds)

}
\seealso{
\code{\link[=getValidScan]{getValidScan()}}
}
