% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{nsnp}
\alias{nsnp}
\alias{nsnp,GbsrGenotypeData-method}
\title{Return the number of SNPs.}
\usage{
nsnp(object, valid = TRUE, ...)

\S4method{nsnp}{GbsrGenotypeData}(object, valid = TRUE)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{valid}{A logical value. See details.}

\item{...}{Unused.}
}
\value{
An integer value to indicate the number of SNP markers.
}
\description{
This function returns the number of SNPs recorded in the GDS file
connected to the given GbsrGenotypeData object.
}
\details{
If \code{valid = TRUE}, the number of SNPs which are labeled \code{TRUE} in
the SnpAnnotationDataSet slot will be returned. You need the number
of over all SNPs, set \code{valid = FALSE}. \code{\link[=getValidSnp]{getValidSnp()}} tells you
which markers are valid.
}
\examples{
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
nsnp(gdata)
closeGDS(gdata) # Close the connection to the GDS file

}
\seealso{
\code{\link[=getValidSnp]{getValidSnp()}}
}
