% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotDosage}
\alias{plotDosage}
\title{Draw line plots of allele dosage per marker per sample.}
\usage{
plotDosage(
  x,
  coord = NULL,
  chr = NULL,
  ind = NULL,
  valid_only = TRUE,
  dot_fill = "blue"
)
}
\arguments{
\item{x}{A GbsrGenotypeData object.}

\item{coord}{A vector with two integer specifying the number of rows and columns to draw faceted line plots for chromosomes.}

\item{chr}{A vector of indexes to specify chromosomes to be drawn.}

\item{ind}{A vector of indexes to specify samples to be drawn.}

\item{valid_only}{A logical value whether to draw a plot only for valid markers and samples. You can get validity with \code{getValidSnp()} and \code{getValidScan}.}

\item{dot_fill}{A string to indicate the dot color in a plot.}
}
\description{
This function counts a reference allele dosage per marker per sample and
draw line plots of them in facets for each chromosome for each sample.
}
