% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{resetFilters}
\alias{resetFilters}
\alias{resetFilters,GbsrGenotypeData-method}
\title{Reset all filters made by \code{\link[=setScanFilter]{setScanFilter()}}, \code{\link[=setSnpFilter]{setSnpFilter()}},
and \code{\link[=setCallFilter]{setCallFilter()}}.}
\usage{
resetFilters(object, ...)

\S4method{resetFilters}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object after removing all filters.

A GbsrGenotypeData object after removing all filters on markers.
}
\description{
Return all data intact.
}
\examples{
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
gdata <- setScanFilter(gdata,
id = getScanID(gdata)[1:10],
missing = 0.2,
dp = c(5, Inf))
gdata <- setSnpFilter(gdata,
 id = getSnpID(gdata)[1:100],
  missing = 0.2,
  dp = c(5, Inf))
gdata <- setInfoFilter(gdata, mq = 40, qd = 20)

# Reset all filters applied above.
gdata <- resetFilters(gdata)

closeGDS(gdata) # Close the connection to the GDS file
}
