% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{resetScanFilters}
\alias{resetScanFilters}
\alias{resetScanFilters,GbsrGenotypeData-method}
\title{Reset the filter made by \code{\link[=setScanFiler]{setScanFiler()}}}
\usage{
resetScanFilters(object, ...)

\S4method{resetScanFilters}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Remove "invalid" labels put on samples and make all samples valid.
}
