% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{resetScanFilters}
\alias{resetScanFilters}
\alias{resetScanFilters,GbsrGenotypeData-method}
\title{Reset the filter made by \code{\link[=setScanFiler]{setScanFiler()}}}
\usage{
resetScanFilters(object, ...)

\S4method{resetScanFilters}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object after removing
all filters on scan(samples).
}
\description{
Remove "invalid" labels put on samples and make all samples valid.
}
\examples{
gds_fn <- system.file("extdata", "simpop.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
gdata <- setScanFilter(gdata,
id = getScanID(gds)[1:10],
missing = 0.2,
dp = c(5, Inf))

# Reset all filters applied above.
gdata <- resetScanFilters(gdata)


}
