% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{resetSnpFilters}
\alias{resetSnpFilters}
\alias{resetSnpFilters,GbsrGenotypeData-method}
\title{Reset the filter made by [setSnpFiler()]}
\usage{
resetSnpFilters(object, ...)

\S4method{resetSnpFilters}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object.
}
\description{
Remove "invalid" labels put on markers and make all markers valid.
}
