% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{resetSnpFilters}
\alias{resetSnpFilters}
\alias{resetSnpFilters,GbsrGenotypeData-method}
\title{Reset the filter made by \code{\link[=setSnpFilter]{setSnpFilter()}}}
\usage{
resetSnpFilters(object, ...)

\S4method{resetSnpFilters}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{...}{Unused.}
}
\value{
A GbsrGenotypeData object after removing all filters on markers.
}
\description{
Remove "invalid" labels put on markers and make all markers valid.
}
\examples{
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
gdata <- loadGDS(gds_fn)
gdata <- setSnpFilter(gdata,
 id = getSnpID(gdata)[1:100],
 missing = 0.2,
 dp = c(5, Inf))

# Reset all filters applied above.
gdata <- resetSnpFilters(gdata)

closeGDS(gdata) # Close the connection to the GDS file
}
