% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{saveScanAnnot}
\alias{saveScanAnnot}
\alias{saveScanAnnot,GbsrGenotypeData-method}
\title{Write out the information stored in the ScanAnnotationDataSet slot}
\usage{
saveScanAnnot(object, ...)

\S4method{saveScanAnnot}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object}

\item{...}{Unused.}
}
\description{
All the data stored in the ScanAnnotationDataSet slot of the GbsrGenotypeData
object can be saved in the GDS file linked to the given GbsrGenotypeData object.
You can load the saved data using \code{\link[=loadSnpAnnot]{loadSnpAnnot()}}.
}
