% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{saveSnpAnnot}
\alias{saveSnpAnnot}
\alias{saveSnpAnnot,GbsrGenotypeData-method}
\title{Write out the information stored in the \link{SnpAnnotationDataFrame} slot}
\usage{
saveSnpAnnot(object, ...)

\S4method{saveSnpAnnot}{GbsrGenotypeData}(object)

\S4method{saveSnpAnnot}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object}

\item{...}{Unused.}
}
\value{
A \link{GbsrGenotypeData} object.
}
\description{
All the data stored in the \code{SnpAnnotatoinDataFrame} slot of
the \link{GbsrGenotypeData}
object can be saved in the GDS file linked to the given
\link{GbsrGenotypeData} object. You can load the saved data
using \code{\link[=loadSnpAnnot]{loadSnpAnnot()}}.
}
\examples{
# Create a GDS file from a sample VCF file.
vcf_fn <- system.file("extdata", "sample.vcf", package = "GBScleanR")
gds_fn <- tempfile("sample", fileext = ".gds")
gbsrVCF2GDS(vcf_fn = vcf_fn, out_fn = gds_fn, force = TRUE)

# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds <- loadGDS(gds_fn)

# Save data stored in the [SnpAnnotationDataFrame] object
# linked in the slot of the `GbsrGenotypeData.`
gds <- saveSnpAnnot(gds)

# Close the connection to the GDS file
closeGDS(gds)

}
