% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{saveSnpAnnot}
\alias{saveSnpAnnot}
\alias{saveSnpAnnot,GbsrGenotypeData-method}
\title{Write out the information stored in the SnpAnnotationDataSet slot}
\usage{
saveSnpAnnot(object, ...)

\S4method{saveSnpAnnot}{GbsrGenotypeData}(object)
}
\arguments{
\item{object}{A GbsrGenotypeData object}

\item{...}{Unused.}
}
\description{
All the data stored in the SnpAnnotatoinDataSet slot of the GbsrGenotypeData
object can be saved in the GDS file linked to the given GbsrGenotypeData object.
You can load the saved data using \code{\link[=loadSnpAnnot]{loadSnpAnnot()}}.
}
