% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{setCallFilter}
\alias{setCallFilter}
\alias{setCallFilter,GbsrGenotypeData-method}
\title{Filter out each genotype call meeting criteria}
\usage{
setCallFilter(
  object,
  dp_count = c(0, Inf),
  ref_count = c(0, Inf),
  alt_count = c(0, Inf),
  norm_dp_count = c(0, Inf),
  norm_ref_count = c(0, Inf),
  norm_alt_count = c(0, Inf),
  scan_dp_qtile = c(0, 1),
  scan_ref_qtile = c(0, 1),
  scan_alt_qtile = c(0, 1),
  snp_dp_qtile = c(0, 1),
  snp_ref_qtile = c(0, 1),
  snp_alt_qtile = c(0, 1),
  omit_geno = NULL,
  ...
)

\S4method{setCallFilter}{GbsrGenotypeData}(
  object,
  dp_count = c(0, Inf),
  ref_count = c(0, Inf),
  alt_count = c(0, Inf),
  norm_dp_count = c(0, Inf),
  norm_ref_count = c(0, Inf),
  norm_alt_count = c(0, Inf),
  scan_dp_qtile = c(0, 1),
  scan_ref_qtile = c(0, 1),
  scan_alt_qtile = c(0, 1),
  snp_dp_qtile = c(0, 1),
  snp_ref_qtile = c(0, 1),
  snp_alt_qtile = c(0, 1),
  omit_geno = NULL
)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{dp_count}{A numeric vector with length two specifying lower and
upper limit of total read counts (reference reads + alternative reads).}

\item{ref_count}{A numeric vector with length two specifying lower and
upper limit of reference read counts.}

\item{alt_count}{A numeric vector with length two specifying lower and
upper limit of alternative read counts.}

\item{norm_dp_count}{A numeric vector with length two specifying lower
and upper limit of normalized total read counts
(normalized reference reads + normalized alternaitve reads).}

\item{norm_ref_count}{A numeric vector with length two specifying lower
and upper limit of normalized reference read counts.}

\item{norm_alt_count}{A numeric vector with length two specifying lower
and upper limit of normalized alternative read counts}

\item{scan_dp_qtile}{A numeric vector with length two specifying lower
and upper limit of quantile of total read counts in each scan (sample).}

\item{scan_ref_qtile}{A numeric vector with length two specifying lower
and upper limit of quantile of reference read counts in each scan (sample).}

\item{scan_alt_qtile}{A numeric vector with length two specifying lower
and upper limit of quantile of alternative read counts in each scan (sample).}

\item{snp_dp_qtile}{A numeric vector with length two specifying lower
and upper limit of quantile of total read counts in each SNP marker}

\item{snp_ref_qtile}{A numeric vector with length two specifying lower
and upper limit of quantile of reference read counts in each SNP marker.}

\item{snp_alt_qtile}{A numeric vector with length two specifying lower
and upper limit of quantile of alternative read counts in each SNP marker.}

\item{omit_geno}{A vector of string with combinations of "ref", "het",
and "alt" to remove specified genotype calls.}

\item{...}{Unused.}
}
\value{
A \link{GbsrGenotypeData} object with filters on genotype calls.
}
\description{
Perform filtering of each genotype call, neither markers
nor samples. Each genotype call is supported by its read counts
for the reference allele and the alternative allele of
a marker of a sample. \code{setCallFilter()} set missing to
the genotype calls which are not reliable enough and set zero
to reference and alternative read counts of the genotype calls.
}
\details{
\code{norm_dp_count}, \code{norm_ref_count}, and \code{norm_alt_count} use
normalized read counts which are obtained by dividing each read count
by the total read count of each sample. \code{scan_dp_qtile}, \code{scan_ref_qtile},
and \code{scan_alt_qtile} work similarly but use quantile values of read counts
of each sample to decide the lower and upper limit of read counts.
This function generate two new nodes in the GDS file linked with
the given \link{GbsrGenotypeData} object. The new nodes "filt.data" in the AD node
and "filt.genotype" contains read count data and genotype data
after filtering, respectively.
To reset the filter applied by setCallFilter(), run \code{\link[=setRawGenotype]{setRawGenotype()}}.
}
\examples{
# Create a GDS file from a sample VCF file.
vcf_fn <- system.file("extdata", "sample.vcf", package = "GBScleanR")
gds_fn <- tempfile("sample", fileext = ".gds")
gbsrVCF2GDS(vcf_fn = vcf_fn, out_fn = gds_fn, force = TRUE)

# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gdata <- loadGDS(gds_fn)

# Filter out genotype calls supported by less than 5 reads.
gdata <- setCallFilter(gdata, dp_count = c(5, Inf))

# Filter out genotype calls supported by reads less than
# the 20 percentile of read counts per marker in each sample.
gdata <- setCallFilter(gdata, scan_dp_qtile = c(0.2, 1))

# Reset the filter
gdata <- setRawGenotype(gdata)

# Close the connection to the GDS file.
closeGDS(gdata)

}
