% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{setFixedParameter}
\alias{setFixedParameter}
\alias{setFixedParameter,GbsrGenotypeData-method}
\title{Set fixed allele read biases and mismapping rate}
\usage{
setFixedParameter(object, bias = NULL, mismap = NULL, parent_geno = FALSE, ...)

\S4method{setFixedParameter}{GbsrGenotypeData}(object, bias, mismap, parent_geno)
}
\arguments{
\item{object}{A \link{GbsrGenotypeData} object.}

\item{bias}{A numeric vector of fixed allele read biases to be assigned to
valid markers. The length of \code{bias} vector should match the number of valid
markers. The values in the \code{bias} vector are assigned to the valid markers
according to their order. NAs in the \code{bias} vector indicates non-fixed biases.}

\item{mismap}{A numeric vector of fixed mismapping rates to be assigned to
valid markers.}

\item{parent_geno}{A matrix of genotypes of parental samples}

\item{...}{Unused.}
}
\value{
A \link{GbsrGenotypeData} object after adding dominant marker information
}
\description{
Set fixed allele read biases of valid markers
}
\details{
Since the bias set by \code{\link[=setFixedParameter]{setFixedParameter()}} function is the reference allele read
bias. Thus, the values 0 and 1 mean that the marker only gives alternative
and reference allele reads, respectively.
Set these fixed biases if some of your markers are dominant markers.
}
\examples{
# Create a GDS file from a sample VCF file.
vcf_fn <- system.file("extdata", "sample.vcf", package = "GBScleanR")
gds_fn <- tempfile("sample", fileext = ".gds")
gbsrVCF2GDS(vcf_fn = vcf_fn, out_fn = gds_fn, force = TRUE)

# Load data in the GDS file and instantiate a [GbsrGenotypeData] object.
gds <- loadGDS(gds_fn)

# Set fixed allele read biases.
# Initialize the bias vector to be assinged.
bias <- rep(NA, nmar(gds))

# As an example, select 20 markers randomly and assign 0 or 1 to them.
# Since the bias set by setFixedParameter() function is the reference allele read
# bias. Thus, the values 0 and 1 means that the marker only gives alternative
# and reference allele reads, respectively.
# Set these fixed biases if some of your markers are dominant markers.
bias[sample(seq_along(bias), 20)] <- sample(c(0, 1), 20, replace = TRUE)

gds <- setFixedParameter(gds, bias = bias)

# Close the connection to the GDS file
closeGDS(gds)

}
