% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Methods-GbsrGenotypeData.R
\name{setValidScan}
\alias{setValidScan}
\alias{setValidScan,GbsrGenotypeData-method}
\title{Manually set valid scans (samples).}
\usage{
setValidScan(object, new, update, ...)

\S4method{setValidScan}{GbsrGenotypeData}(object, new, update)
}
\arguments{
\item{object}{A GbsrGenotypeData object.}

\item{new}{A logical vector of the same length with the over all number of the samples.}

\item{update}{A logical vector of the same length with the currently valid samples.}

\item{...}{Unused.}
}
\description{
If you need manually set valid and invalid samples, you can do it via this function,
e.g in the case you conducted a filtering on samples manually by your self.
}
\details{
To over write the current validity information, give a logical vector to `new`.
On the other hand, a logical vector specified to `update` will be used to
update validity information of the currently valid samples. If you gave
a vector for both argument, only the vector passed to `new` will be used to
over write the validity information.
}
\seealso{
[setScanFilter()] to filter out samples based on some summary statistics
}
